chore: include in new page layout
