--- src/main/resources/templates/auth-flow.html
@@ -1,10 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Digg Wallet Sandbox</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-    <h1>Auth flow</h1>
-</body>
-</html>
