+++ src/main/resources/templates/credential-offer.html
@@ -1,15 +1,26 @@
 <!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Digg Wallet Sandbox</title>
-        <link href="./style/digg.css" rel="stylesheet" />
-    </head>
-    <body>
-        <img th:if="${qrCode != null and !#strings.isEmpty(qrCode)}" th:src="${qrCode}" alt="scan in wallet app in order to issue personal id">
-        <a th:href="${linkUrl}">Hämta till plånbok</a>
-    </body>
-</html>
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
 
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Digg Wallet Sandbox</title>
+    <link th:href="@{./style/digg.css}" rel="stylesheet" />
+    <meta name="description"
+          content="DIGG Wallet Sandbox for EUDIW Reference Implementation and Personal Identity Document (PID) Issuance">
+</head>
+<header th:replace="~{layout :: header}"></header>
+
+<body>
+<main class="main-container">
+    <div class="content-area">
+            <img th:if="${qrCode != null and !#strings.isEmpty(qrCode)}" th:src="${qrCode}" alt="scan in wallet app in order to issue personal id">
+            <a th:href="${linkUrl}">Hämta till plånbok</a>
+          </div>
+        </main>
+        <script th:src="@{/js/script.js}"></script>
+    </body>
+    <footer th:replace="~{layout :: footer}"></footer>
+
+</html>
