+++ src/main/resources/templates/index.html
@@ -104,7 +104,7 @@
                         <h4 class="article-title">What can I test in the apps?</h4>
 
                         <p class="article-content">You can apply for a PID from Digg either by scanning a QR-code for a
-                            <a href="./pid">credential offer</a> or selecting a
+                            <a href="./prepare-credential-offer">credential offer</a> or selecting a
                             PID from the preconfigured list you can find in the app. You can also use your newly
                             acquired PID and have it verified
                             on <a href="https://wallet.sandbox.digg.se/verifier/">Diggs Verifier</a> either by selective
