--- src/main/resources/templates/pid-credential-offer.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Digg Wallet Sandbox</title>
-    <link href="./style/digg.css" rel="stylesheet" />
-</head>
-<body>
-<div th:replace="~{layout :: header}"></div>
-<main class="main-container">
-    <div class="content-area">
-        <section class="content top-stories">
-            <div class="section-header">Create PID credential offer</div>
-
-            <article>
-                <div class="article-image">
-                    🪪
-                </div>
-                <div style="margin: 30px;">
-                    <img th:if="${qrCode != null and !#strings.isEmpty(qrCode)}" th:src="${qrCode}" alt="scan in wallet app in order to issue personal id">
-                    <form action="#" th:action="@{/pid}" th:object="${credentialOffer}" method="post">
-                        <p>Pid Mso Mdoc: <input type="checkbox" th:field="*{pidMsoMdoc}" /></p>
-                        <p>Pid Sd Jwt Vc: <input type="checkbox" th:field="*{pidSdJwtVc}" /></p>
-                        <input type="hidden" th:field="*{preAuthCodeFlow}" />
-                        <p th:text="*{validationErrors}"></p>
-                        <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
-                    </form>
-                    <a th:href="${linkUrl}">Fetch to wallet</a>
-                </div>
-            </article>
-        </section>
-    </div>
-</main>
-<div th:replace="~{layout :: footer}"></div>
-</body>
-</html>
-
