--- src/main/resources/templates/pid-pre-auth-credential-offer.html
@@ -1,115 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Digg Wallet Sandbox</title>
-    <link href="../style/digg.css" rel="stylesheet" />
-</head>
-<div th:replace="~{layout :: header}"></div>
-<body>
-<header>
-    <div class="header-container">
-        <div class="logo">
-            <a href="https://www.digg.se" aria-label="Gå till digg.se">
-                <img alt="DIGG - Myndigheten för digital förvaltning" class="sv-noborder sv-svg" src="../images/DIGG-logo-color.svg">
-            </a>
-        </div>
-        <nav>
-            <ul>
-                <li><a href="#">EU Digital Identity Wallet</a></li>
-            </ul>
-        </nav>
-    </div>
-</header>
-
-<main class="main-container">
-    <div class="content-area">
-        <section class="content top-stories">
-            <div class="section-header">Personal Identifier (PID) Pre-Auth Credential offer</div>
-
-            <article class="article">
-                <div class="article-image">
-                    🪪
-                </div>
-                <div class="article-content">
-    <img th:src="${qrCode}" alt="scan in wallet app in order to issue personal id">
-    <a th:href="${linkUrl}">Hämta till plånbok</a>
-    <h2>PIN code</h2>
-    <form>
-        <label for="txCode">txCode</label><input type="text" readonly id="txCode" th:value="${txCode}"/>
-        <div style="display: inline-block;
-    padding: 6px 12px;
-    margin-bottom: 0;
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 1.42857143;
-    text-align: center;
-    white-space: nowrap;
-    vertical-align: middle;
-    -ms-touch-action: manipulation;
-    touch-action: manipulation;
-    cursor: pointer;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    background-image: none;
-    border: 1px solid transparent;
-    border-radius: 4px;
-    text-decoration: none;
-
-    color: #333;
-    background-color: #fff;
-    border-color: #ccc;" onclick="copyTxCode()">Copy txCode</div>
-        <button>Generera ny credential</button>
-    </form>
-
-                </div>
-            </article>
-        </section>
-
-    </div>
-
-</main>
-
-
-<footer>
-    <div class="footer-container">
-        <div class="footer-links">
-            <div>
-                <a href="https://www.linkedin.com/company/digg-sverige/" class="dg-normal dg-external-link  dg-footer-social-link LinkedIn" lang="en"><span class="env-assistive-text">DIGG på&nbsp;</span><span>LinkedIn</span></a>
-                <a href="https://press.newsmachine.com/pressroom/view/digg-myndigheten-for-digital-forvaltning" class="dg-normal dg-external-link  dg-footer-social-link PressMachine" lang="en"><span class="env-assistive-text">DIGG på&nbsp;</span><span>PressMachine</span></a>
-            </div>
-            <div>
-                <a href="mailto:info@digg.se">info@digg.se</a>
-                <a href="tel:0771-114400">Tel: 0771-11 44 00</a>
-            </div>
-
-        </div>
-        <div class="copyright">© 2025 <strong>Digg</strong> - Myndigheten för digital förvaltning. All rights reserved.</div>
-    </div>
-</footer>
-</body>
-<<<<<<< Updated upstream
-
-<script>
-    function copyTxCode() {
-        // Get the text field
-        var copyText = document.getElementById("txCode");
-
-        // Select the text field
-        copyText.select();
-        copyText.setSelectionRange(0, 99999); // For mobile devices
-
-        // Copy the text inside the text field
-        navigator.clipboard.writeText(copyText.value);
-    }
-</script>
-
-||||||| Stash base
-=======
-<div th:replace="~{layout :: footer}"></div>
->>>>>>> Stashed changes
-</html>
-
