+++ src/main/resources/templates/prepare-credential-offer.html
@@ -1,101 +1,108 @@
 <!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Digg Wallet Sandbox</title>
-        <link href="./style/digg.css" rel="stylesheet" />
-        <style>
-            /* Hide the default checkbox */
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
 
-            .custom-checkbox {
-                margin: 2px;
-                padding: 2px;
-            }
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Digg Wallet Sandbox</title>
+    <link th:href="@{./style/digg.css}" rel="stylesheet" />
+    <style>
+        /* Hide the default checkbox */
 
-            .checkbox-input {
-                position: absolute;
-                opacity: 0;
-                cursor: pointer;
-                height: 0;
-                width: 0;
-            }
+        .custom-checkbox {
+            margin: 2px;
+            padding: 2px;
+        }
 
-            /* Style the label container */
-            .checkbox-label {
-                display: flex;
-                cursor: pointer;
-                user-select: none;
-                border: 2px solid #ccc;
-                border-radius: 6px; /* Rounded corners */
-                box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2); /* Shadow effect */
-                transition: all 0.2s ease;
-                margin: 5px;
-                padding: 10px;
+        .checkbox-input {
+            position: absolute;
+            opacity: 0;
+            cursor: pointer;
+            height: 0;
+            width: 0;
+        }
 
-            }
+        /* Style the label container */
+        .checkbox-label {
+            display: flex;
+            cursor: pointer;
+            user-select: none;
+            border: 2px solid #ccc;
+            border-radius: 6px; /* Rounded corners */
+            box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2); /* Shadow effect */
+            transition: all 0.2s ease;
+            margin: 5px;
+            padding: 10px;
 
-            /* Style the custom checkbox with rounded corners and shadow */
-            .checkbox-custom {
-                display: flex;
-                flex-direction: column;
-                justify-content: flex-start;
-                width: 20px;
-                height: 20px;
-                border: 1px solid #ccc;
-                border-radius: 6px; /* Rounded corners */
-                box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2); /* Shadow effect */
-                transition: all 0.2s ease;
-                background-color: #fff; /* Default background color */
-            }
+        }
 
-            .checkbox-icon {
-                width: 16px;
-                height: 16px;
-                fill: white;
-                visibility: hidden;
-            }
+        /* Style the custom checkbox with rounded corners and shadow */
+        .checkbox-custom {
+            display: flex;
+            flex-direction: column;
+            justify-content: flex-start;
+            width: 20px;
+            height: 20px;
+            border: 1px solid #ccc;
+            border-radius: 6px; /* Rounded corners */
+            box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2); /* Shadow effect */
+            transition: all 0.2s ease;
+            background-color: #fff; /* Default background color */
+        }
 
-            /* Style for the icon inside the checkbox */
-            .checkbox-label-icon {
-                width: 48px;
-                height: 48px;
-                fill: white;
-                margin: 5px 10px 5px 20px;
-            }
+        .checkbox-icon {
+            width: 16px;
+            height: 16px;
+            fill: white;
+            visibility: hidden;
+        }
 
-            /* Style for when checkbox is checked - different background color */
-            .checkbox-input:checked ~ .checkbox-label .checkbox-custom {
-                background-color: #4A90E2; /* Blue background when checked */
-                border-color: #3A80D2;
-                box-shadow: 0 2px 6px rgba(74, 144, 226, 0.4); /* Enhanced shadow when checked */
-            }
+        /* Style for the icon inside the checkbox */
+        .checkbox-label-icon {
+            width: 48px;
+            height: 48px;
+            fill: white;
+            margin: 5px 10px 5px 20px;
+        }
 
-            .checkbox-input:checked ~ .checkbox-label .checkbox-icon {
-                visibility: visible;
-            }
+        /* Style for when checkbox is checked - different background color */
+        .checkbox-input:checked ~ .checkbox-label .checkbox-custom {
+            background-color: #4A90E2; /* Blue background when checked */
+            border-color: #3A80D2;
+            box-shadow: 0 2px 6px rgba(74, 144, 226, 0.4); /* Enhanced shadow when checked */
+        }
 
-            .checkbox-input:checked ~ .checkbox-label {
-                background-color: #d5dad4;
-            }
+        .checkbox-input:checked ~ .checkbox-label .checkbox-icon {
+            visibility: visible;
+        }
 
-            /* Focus styles for accessibility */
-            .checkbox-input:focus ~ .checkbox-label .checkbox-custom {
-                box-shadow: 0 0 0 3px rgba(74, 144, 226, 0.3);
-            }
+        .checkbox-input:checked ~ .checkbox-label {
+            background-color: #d5dad4;
+        }
 
-            /* Style the text */
-            .checkbox-text {
-                font-size: 18px;
-                align-self: center;
-            }
+        /* Focus styles for accessibility */
+        .checkbox-input:focus ~ .checkbox-label .checkbox-custom {
+            box-shadow: 0 0 0 3px rgba(74, 144, 226, 0.3);
+        }
 
-            .auth-flow-select {
+        /* Style the text */
+        .checkbox-text {
+            font-size: 18px;
+            align-self: center;
+        }
 
-            }
-        </style>
-    </head>
-    <body>
+        .auth-flow-select {
+
+        }
+    </style>
+    <meta name="description"
+          content="DIGG Wallet Sandbox for EUDIW Reference Implementation and Personal Identity Document (PID) Issuance">
+</head>
+<header th:replace="~{layout :: header}"></header>
+
+<body>
+<main class="main-container">
+    <div class="content-area">
         <form action="" th:action="@{/prepare-credential-offer}" th:object="${credentialOffer}" method="post">
             <div th:each="item : ${credentialOffer.availableCredentials}">
                 <div class="custom-checkbox">
@@ -121,7 +128,11 @@
 
             <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
         </form>
-    </body>
-</html>
+    </div>
+</main>
+<script th:src="@{/js/script.js}"></script>
+</body>
+<footer th:replace="~{layout :: footer}"></footer>
 
+</html>
